// src/domain/common/errors/enums/EErrorCode.ts
export enum EErrorCode {
  ADDRESS_INVALID_NUMBER = 'ADDRESS_INVALID_NUMBER',
  ADDRESS_INVALID_STATE = 'ADDRESS_INVALID_STATE',
  ADDRESS_INVALID_ZIP_CODE = 'ADDRESS_INVALID_ZIP_CODE',
  ADVANCE_STAGE_MISSING_FIELDS = 'ADVANCE_STAGE_MISSING_FIELDS',
  AUTH_INVALID_CREDENTIALS = 'AUTH_INVALID_CREDENTIALS',
  AUTH_UNAUTHORIZED = 'AUTH_UNAUTHORIZED',
  CANNOT_DISMISS_WARNING = 'CANNOT_DISMISS_WARNING',
  CHAT_CONTENT_REJECTED = 'CHAT_CONTENT_REJECTED',
  CHAT_CONVERSATION_BLOCKED = 'CHAT_CONVERSATION_BLOCKED',
  CHAT_NOT_ELIGIBLE = 'CHAT_NOT_ELIGIBLE',
  COGNITO_USER_ALREADY_CONFIRMED = 'COGNITO_USER_ALREADY_CONFIRMED',
  COGNITO_USER_CONFIRM_FAILED = 'COGNITO_USER_CONFIRM_FAILED',
  COGNITO_USER_NOT_FOUND_BY_EMAIL = 'COGNITO_USER_NOT_FOUND_BY_EMAIL',
  COGNITO_USER_NOT_FOUND_BY_SUB = 'COGNITO_USER_NOT_FOUND_BY_SUB',
  DATABASE_ERROR = 'DATABASE_ERROR',
  FIELD_INVALID = 'FIELD_INVALID',
  FIELD_MAX_LENGTH = 'FIELD_MAX_LENGTH',
  FIELD_MAX_VALUE = 'FIELD_MAX_VALUE',
  FIELD_MIN_LENGTH = 'FIELD_MIN_LENGTH',
  FIELD_MIN_VALUE = 'FIELD_MIN_VALUE',
  INVALID_PAGE_SIZE = 'INVALID_PAGE_SIZE',
  LISTING_ALREADY_RESERVED = 'LISTING_ALREADY_RESERVED',
  LISTING_NOT_AVAILABLE_FOR_PURCHASE = 'LISTING_NOT_AVAILABLE_FOR_PURCHASE',
  MAPS_ERROR = 'MAPS_ERROR',
  NATIONAL_ID_INVALID = 'NATIONAL_ID_INVALID',
  NATIONAL_ID_INVALID_LENGTH = 'NATIONAL_ID_INVALID_LENGTH',
  PAYMENT_INSUFFICIENT_FUNDS = 'PAYMENT_INSUFFICIENT_FUNDS',
  PROOF_OF_RESIDENCE_ADDRESS_MISSING = 'PROOF_OF_RESIDENCE_ADDRESS_MISSING',
  REJECTION_REASON_DESC_TOO_SHORT = 'REJECTION_REASON_DESC_TOO_SHORT',
  REJECTION_REASON_NOT_ALLOWED = 'REJECTION_REASON_NOT_ALLOWED',
  REJECTION_REASON_REQUIRED = 'REJECTION_REASON_REQUIRED',
  RESOURCE_CONFLICT = 'RESOURCE_CONFLICT',
  RESOURCE_NOT_FOUND = 'RESOURCE_NOT_FOUND',
  SERVICE_DEPENDENT_INACTIVE = 'SERVICE_DEPENDENT_INACTIVE',
  SERVICE_USER_UNKNOWN = 'SERVICE_USER_UNKNOWN',
  STATUS_REQUIRES_FIELDS = 'STATUS_REQUIRES_FIELDS',
  STATUS_REQUIRES_PROOF_ADDRESS = 'STATUS_REQUIRES_PROOF_ADDRESS',
  STATUS_REQUIRES_PROOF_FILE = 'STATUS_REQUIRES_PROOF_FILE',
  UNDER_REVIEW_REQUIRES_WAITING_REVIEW = 'UNDER_REVIEW_REQUIRES_WAITING_REVIEW',
  USER_CREATE_FAILED = 'USER_CREATE_FAILED',
  USER_DELETE_FAILED = 'USER_DELETE_FAILED',
  USER_DISMISS_WARNING_FAILED = 'USER_DISMISS_WARNING_FAILED',
  USER_FIND_FAILED = 'USER_FIND_FAILED',
  USER_RESET_FAILED = 'USER_RESET_FAILED',
  USER_UNDERAGE = 'USER_UNDERAGE',
  USER_UPDATE_FAILED = 'USER_UPDATE_FAILED',
  WAITING_REVIEW_DATA_INCOMPLETE = 'WAITING_REVIEW_DATA_INCOMPLETE',
}